mudanca

+++ Rest/src/java/controller/resources/UserResource.java
@@ -2,6 +2,7 @@ package controller.resources;
 
 import com.google.gson.Gson;
 import java.util.List;
+import javax.ws.rs.Consumes;
 
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
@@ -21,12 +22,12 @@ import model.vo.User;
 @Path("/user")
 public class UserResource {
 
-    private UserDAO userDAO = new UserDAO();
 
     @GET
     @Path("/getUsers")
     @Produces(MediaType.APPLICATION_JSON)
     public Response getUsers() {
+        UserDAO userDAO = new UserDAO();
         List<User> list = userDAO.listAll();
         if (list.size() == 0) {
             return Response.status(Response.Status.NOT_FOUND).build();
@@ -37,7 +38,8 @@ public class UserResource {
     @GET
     @Path("/getUser/{login}")
     @Produces(MediaType.APPLICATION_JSON)    
-    public Response getUser(@PathParam("login") String login) {   
+    public Response getUser(@PathParam("login") String login) {
+        UserDAO userDAO = new UserDAO();
         User user = userDAO.getUser(login);
         if (user == null) {
             return Response.status(Response.Status.NOT_FOUND).build();
@@ -46,26 +48,42 @@ public class UserResource {
     }
 
     @POST
-    @Path("/addUser/{login}/{password}")
-    public Response addUser(@PathParam("login") String login, @PathParam("password") String password) {
+    @Path("/addUser")
+    @Consumes(MediaType.APPLICATION_JSON)
+    public Response addUser(String json) {
+        User user = new Gson().fromJson(json, User.class);
+        if (user.getLogin() == null || user.getPassword() == null || user.getLogin().equals("") || user.getPassword().equals("")) {
+            return Response.status(Response.Status.BAD_REQUEST).build();
+        }
+        UserDAO userDAO = new UserDAO();
+        userDAO.add(user);
+        return Response.status(Response.Status.CREATED).build();
+    }
+   /*
+    @PUT
+    @Path("/updateUser/{login}")
+    public Response addUser(@BodyParam body, @PathParam("login") String login, @PathParam("password") String password) {
         if (login == null || password == null || login.equals("") || password.equals("")) {
             return Response.status(Response.Status.BAD_REQUEST).build();
         }
         User user = new User();
         user.setLogin(login);
-        user.setPassword(password);        
+        user.setPassword(password);
         userDAO.add(user);
         return Response.status(Response.Status.CREATED).build();
     }
-
+*/
     @DELETE
     @Path("/deleteUser/{login}")
-    public Response deleteUser(@PathParam("login") String login) {
+    public Response deleteUser(@PathParam("login") String login) {       
         if (login == null || login.equals("")) {
             return Response.status(Response.Status.BAD_REQUEST).build();
         }
-        User user = new User();
-        user.setLogin(login);
+        UserDAO userDAO = new UserDAO();
+        User user = userDAO.getUser(login);
+        if (user == null) {
+            return Response.status(Response.Status.NOT_FOUND).build();
+        }
         userDAO.delete(user);
         return Response.status(Response.Status.OK).build();
     }
